Refactoring: split classes.py into submodules. Change naming
